Moved Aggregation.kt to core
This commit moves Aggregation.kt to core, to
resolve a cyclic dependency.

See: gh-1024
